move components from common to src
