+++ packages/frinx-workflow-ui/src/components/workflow-nodes/base-node.tsx
@@ -1,5 +1,5 @@
 import { Box, Heading, Text } from '@chakra-ui/react';
-import { ExtendedTask } from '@frinx/shared/src';
+import { ExtendedTask } from 'packages/shared/src';
 import React, { memo, VoidFunctionComponent } from 'react';
 import { Handle, NodeProps, Position } from 'react-flow-renderer';
 import { ExecutionState, getBackgroundColor } from './node.helpers';
