+++ packages/frinx-workflow-ui/src/components/workflow-nodes/node.helpers.ts
